use C++ for peak_memusage wrapper
